Add support for webhooks from Slack
Slack incoming webhooks are form-urlencoded instead of application/json.  This adds an interceptor that can take in the form-urlencoded stuff from slack and make it look like application json (prepended with slack. see Triggerbindings)

This adds event listeners and bindings to take a number of key elements out of an incoming slash command payload