Attempt to address volume node affinity conflict issues
The seed cache task that runs during installation now requests both workspaces to be used in the pipeline to ensure they both get allocated to the same az.

This should in theory avoid the repo pvc getting allocated to one az and then later a task in the pipeline run (e.g. git-clone) allocating the workspace pvc to a separate az